All materials for Distance Instrument completed
A lot to work on here. Definitely missing some important functionality with distance sensor based on how my metro is currently responding. Definitely want to get your thoughts on that as I may have explained it poorly but its shown in the video when I'm making very drastic changes to the distance